Refactor boundary codec ownership
