+++ apps/server/tests/integration/test_contract_analysis_persistence_http.py
@@ -13,15 +13,14 @@ from vibesensor.adapters.history import ProjectedHistoryRunService
 from vibesensor.domain import RunStatus
 from vibesensor.shared.boundaries.persisted_analysis_codec import (
     persisted_analysis_from_storage_json_object,
-    persisted_analysis_from_summary,
     persisted_analysis_to_storage_json_object,
-    persisted_analysis_to_summary,
 )
 from vibesensor.shared.boundaries.run_metadata_codec import run_metadata_from_mapping
 from vibesensor.shared.types.history_analysis_contracts import AnalysisSummary
 from vibesensor.shared.types.history_records import StoredHistoryRun
 from vibesensor.shared.types.persisted_analysis import (
     PERSISTED_ANALYSIS_SCHEMA_VERSION,
+    PersistedAnalysis,
 )
 from vibesensor.use_cases.history.runs import HistoryRunService
 
@@ -64,12 +63,12 @@ def _representative_summary() -> AnalysisSummary:
 def _stored_run_from_summary(
     summary: AnalysisSummary,
 ) -> tuple[StoredHistoryRun, AnalysisSummary]:
-    persisted = persisted_analysis_from_summary(summary)
+    persisted = PersistedAnalysis.from_json_object(summary)
     storage_payload = persisted_analysis_to_storage_json_object(persisted)
     assert storage_payload["_schema_version"] == PERSISTED_ANALYSIS_SCHEMA_VERSION
 
     reloaded = persisted_analysis_from_storage_json_object(storage_payload)
-    restored_summary = persisted_analysis_to_summary(reloaded)
+    restored_summary = cast(AnalysisSummary, reloaded.to_json_object())
 
     metadata = run_metadata_from_mapping(
         {
