+++ apps/server/tests/shared/boundaries/test_analysis_summary_projection.py
@@ -12,9 +12,7 @@ from vibesensor.shared.boundaries.analysis_summary_projection import (
     project_persisted_analysis,
 )
 from vibesensor.shared.boundaries.finding import finding_payload_from_domain
-from vibesensor.shared.boundaries.persisted_analysis_codec import (
-    persisted_analysis_from_json_object,
-)
+from vibesensor.shared.types.persisted_analysis import PersistedAnalysis
 
 
 def _canonical_metadata() -> dict[str, object]:
@@ -157,7 +155,7 @@ def test_project_persisted_analysis_uses_persisted_reconstruction_path(
 ) -> None:
     import vibesensor.shared.boundaries.analysis_summary_projection as projection
 
-    analysis = persisted_analysis_from_json_object(
+    analysis = PersistedAnalysis.from_json_object(
         {
             "case_id": "case-001",
             "run_id": "run-001",
